pyproject.toml
[tool.poetry]
name = "blob-effect"
version = "v1.1.2"
description = ""
authors = ["Kodai YAMASHITA <nglcobdai@gmail.com>"]
readme = "README.md"
packages = [
    { include = "nglcobdai_utils", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
numpy = "^1.26.0"
pillow = "^10.1.1"
opencv-python = "^4.8.1.78"
pyyaml = "^6.0.1"
tqdm = "^4.66.1"
matplotlib = "^3.8.1"
nglcobdai-utils = {git = "https://github.com/nglcobdai/nglcobdai-utils.git", tag = "v0.2.0" }


[tool.poetry.group.dev.dependencies]
pep8 = "^1.7.1"
black = "^24.3.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 120
